Restructured election tests
