+++ tests/aswwu/behaviors/elections/candidate/candidate_requests.py
@@ -1,0 +1,5 @@
+# (r"/elections/election/(.*)/candidate", elections.CandidateHandler)
+# get, post
+
+# (r"/elections/election/(.*)/candidate/(.*)", elections.SpecifiedCandidateHandler)
+# get, put, delete
