+++ tests/aswwu/behaviors/elections/election/election_subtests.py
@@ -1,13 +1,13 @@
 import json
-import tests.aswwu.behaviors.elections.elections_requests as elections_requests
+import tests.aswwu.behaviors.elections.election.election_requests as election_requests
 from tests.utils import load_csv
-from tests.aswwu.data.paths import ELECTIONS_PATH, POSITIONS_PATH
+from tests.aswwu.data.paths import ELECTIONS_PATH
 
 from datetime import datetime, timedelta
 
 
 def send_get_election():
-    resp = elections_requests.get_election()
+    resp = election_requests.get_election()
     assert (resp.status_code == 200)
     resp_text = json.loads(resp.text)
     elections_csv = load_csv(ELECTIONS_PATH)
@@ -22,8 +22,8 @@ def send_get_election():
 
 
 def _post_election(election):
-    resp = elections_requests.post_election(election['election_type'], election['name'], election['max_votes'],
-                                            election['start'], election['end'], election['show_results'])
+    resp = election_requests.post_election(election['election_type'], election['name'], election['max_votes'],
+                                           election['start'], election['end'], election['show_results'])
     resp_text = json.loads(resp.text)
     assert (resp.status_code == 201)
     assert (resp_text['election_type'] == election['election_type'])
@@ -32,12 +32,16 @@ def _post_election(election):
     assert (resp_text['start'] == election['start'])
     assert (resp_text['end'] == election['end'])
     assert (resp_text['show_results'] == election['show_results'])
+    return resp_text
 
 
 def send_post_election():
+    elections_data = {}
     elections = load_csv(ELECTIONS_PATH)
     for election in elections:
-        _post_election(election)
+        data = _post_election(election)
+        elections_data[data['id']] = data
+    return elections_data
 
 
 def send_post_dynamic_election():
@@ -46,35 +50,15 @@ def send_post_dynamic_election():
         'election_type': 'aswwu',
         'name': 'General Election Test',
         'max_votes': 2,
-        'start': datetime.now(),
-        'end': datetime.now() + timedelta(days=1),
-        'show_results': datetime.now() + timedelta(days=1),
+        'start': datetime.strftime(datetime.now() + timedelta(hours=1), '%Y-%m-%d %H:%M:%S.%f'),
+        'end': datetime.strftime(datetime.now() + timedelta(days=1), '%Y-%m-%d %H:%M:%S.%f'),
+        'show_results': datetime.strftime(datetime.now() + timedelta(days=1), '%Y-%m-%d %H:%M:%S.%f'),
     }
     _post_election(new_election)
 
 
 def send_get_current():
-    resp = elections_requests.get_current()
+    resp = election_requests.get_current()
     resp_text = json.loads(resp.text)
     assert(resp.status_code == 200)
     assert(datetime.strptime(resp_text['end'], "%Y-%m-%d %H:%M:%S.%f") >= datetime.now())
-
-
-def send_get_position():
-    resp = elections_requests.get_position()
-    resp_text = json.loads(resp.text)
-    assert(resp.status_code == 200)
-
-
-def send_post_position():
-    positions = load_csv(POSITIONS_PATH)
-    for position in positions:
-        resp = elections_requests.post_position(position['position'], position['election_type'], position['active'],
-                                                position['order'])
-        resp_text = json.loads(resp.text)
-        assert (resp.status_code == 201)
-        assert(resp_text['position'] == position['position'])
-        assert(resp_text['election_type'] == position['election_type'])
-        assert(str(resp_text['active']) == position['active'])
-        assert(str(resp_text['order']) == position['order'])
-
