+++ tests/aswwu/behaviors/elections/election/test_election.py
@@ -1,4 +1,4 @@
-from tests.aswwu.behaviors.elections.elections_subtests import *
+from tests.aswwu.behaviors.elections.election.election_subtests import *
 from tests.conftest import testing_server
 
 
@@ -11,13 +11,5 @@ def test_get_election(testing_server):
 
 
 def test_get_current(testing_server):
-    send_get_current()
-
-
-def test_post_position(testing_server):
-    send_post_position()
-
-
-def test_get_position(testing_server):
     send_post_dynamic_election()
-    send_get_position()
+    send_get_current()
