+++ tests/aswwu/behaviors/elections/vote/vote_subtests.py
@@ -1,0 +1,2 @@
+def send_post_vote():
+    votes = load_csv(VOTES_PATH)
