+++ tests/aswwu/data/paths.py
@@ -2,3 +2,4 @@ __data_path = 'tests/aswwu/data/'
 USERS_PATH = __data_path + 'users.csv'
 ELECTIONS_PATH = __data_path + 'elections.csv'
 POSITIONS_PATH = __data_path + 'positions.csv'
+VOTES_PATH = __data_path + 'votes.csv'
