+++ tests/aswwu/data/votes.csv
@@ -1,0 +1,10 @@
+vote
+armanda.Woolston
+Holly.garza
+raeann.Castor
+tammie.treacy
+Virgilio.mccraw
+Lashay.Semien
+delcie.Lauer
+Melva.woullard
+Eugene.burnette
